+++ feature_audiobook_enhance_details/src/main/java/com/allsoftdroid/audiobook/feature/feature_audiobook_enhance_details/data/repository/FetchAdditionalBookDetailsRepositoryImpl.kt
@@ -6,13 +6,15 @@ import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.data
 import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.domain.repository.IFetchAdditionBookDetailsRepository
 import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.domain.network.NetworkResponseListener
 import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.domain.repository.IStoreRepository
+import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.utils.BookDetailsParserFromHtml
 import com.dropbox.android.external.store4.get
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import timber.log.Timber
 
 class FetchAdditionalBookDetailsRepositoryImpl(private val storeCachingRepository: IStoreRepository,
-                                               private val bookDetailsParser: BookDetailsParserFromHtml) :
+                                               private val bookDetailsParser: BookDetailsParserFromHtml
+) :
     IFetchAdditionBookDetailsRepository {
 
     private var _bookDetails = MutableLiveData<BookDetails>()
